# docs/openapi.yaml
openapi: 3.0.3
info:
  title: J-J ERP API
  version: "1.0.0"
  description: >
    API REST del sistema ERP J-J. Desarrollada con **Laravel 12** y
    autenticada mediante **Laravel Sanctum** (Bearer Token).
    Permite gestionar productos, ventas, facturas, empleados, clientes,
    jornadas laborales y más.
  contact:
    email: admin@negocio.test
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Desarrollo local
  - url: https://api.negocio.test
    description: Producción AWS
security:
  - sanctum: []
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: Token Sanctum obtenido en POST /api/login
  schemas:
    Producto:
      type: object
      properties:
        id:             { type: integer, example: 1 }
        nombre:         { type: string,  example: "Café Espresso" }
        sku:            { type: string,  example: "CAF-001" }
        descripcion:    { type: string,  example: "Café de especialidad" }
        precio:         { type: number,  format: float, example: 2.50 }
        stock_quantity: { type: integer, example: 100 }
        categoria_id:   { type: integer, example: 3 }
        proveedor_id:   { type: integer, example: 2 }
...
    User:
      type: object
      properties:
        id:     { type: integer }
        nombre: { type: string }
        email:  { type: string, format: email }
        rol:    { type: string, enum: [admin, gerente, vendedor] }
...
paths:
...
  /api/login:
    post:
      tags: [Autenticación]
      summary: Iniciar sesión y obtener token Sanctum
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email, example: "admin@negocio.test" }
                password: { type: string, format: password, example: "password" }
      responses:
        '200':
          description: Login correcto
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, example: "1|abc123xyz..." }
                  user:  { $ref: '#/components/schemas/User' }
        '401': { description: Credenciales inválidas }
        '422': { description: Datos inválidos }
  /api/logout:
    post:
      tags: [Autenticación]
      summary: Cerrar sesión (revocar token actual)
      responses:
        '200': { description: Sesión cerrada }
        '401': { description: No autenticado }

  # ────────────────────────────────────────────────
  # PRODUCTOS
  # ────────────────────────────────────────────────
  /api/productos:
    get:
      tags: [Productos]
      summary: Listar todos los productos
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Producto' }
        '401': { description: No autenticado }
    post:
      tags: [Productos]
      summary: Crear un nuevo producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Producto'
      responses:
        '201': { description: Producto creado }
        '422': { description: Datos inválidos }
        '401': { description: No autenticado }
  /api/productos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Productos]
      summary: Obtener producto por ID
      responses:
        '200':
          description: Producto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Producto' }
        '404': { description: No encontrado }
    put:
      tags: [Productos]
      summary: Actualizar producto
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Producto' }
      responses:
        '200': { description: Actualizado }
        '404': { description: No encontrado }
    delete:
      tags: [Productos]
      summary: Eliminar producto
      responses:
        '200': { description: Eliminado }
        '404': { description: No encontrado }

  # ────────────────────────────────────────────────
  # CLIENTES
  # ────────────────────────────────────────────────
...